python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "management-helper"
version = "0.1.0"
description = "仕入管理・ラベル印刷自動化ツール"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "gspread==6.1.2",
    "oauth2client==4.1.3",
    "google-api-python-client>=2.0",
    "httpx>=0.27",
    "python-dotenv==1.0.1",
    "click>=8.0",
    "tenacity>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest==8.3.3",
    "pytest-mock==3.14.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
pythonpath = ["src"]
